og_requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-language==2.9.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
grpcio-tools==1.53.0
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
nltk==3.8.1
pip-autoremove==0.10.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
textblob==0.17.1
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
